NLPBACK/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
numpy==1.25.2
sentence-transformers==2.2.2
scikit-learn==1.3.0
python-multipart==0.0.6
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0
python-dotenv==1.0.0
huggingface-hub==0.16.4 